Huge refactoring of files to make things clearer. updated package.json, now need app to load over the index.html page
