+++ client/html/index.html
@@ -1,0 +1,13 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+  <title>WotWot Rendered server-side</title>
+
+  <link rel="stylesheet" href="./main.css">
+</head>
+<body>
+  <p> loading screen is gangsta</p>
+  <div id="app"></div>
+  <script src="bundle.js"></script>
+</body>
+</html>
