+++ package.json
@@ -1,14 +1,16 @@
 {
   "name": "wotwot-v2",
   "version": "1.0.0",
-  "description": "A clean restart for WotWot",
+  "description": "This is the final restart",
   "main": "server/server.js",
   "scripts": {
-    "build": "webpack",
-    "watch": "webpack --watch",
-    "start": "webpack-dev-server --hot --inline",
-    "server": "node-inspector --web-port 8888 & node-dev server",
-    "test": "ava --watch"
+    "dev-server": "webpack-dev-server",
+    "serve": "webpack-dev-server --hot --inline",
+    "build": "webpack && npm run transpile",
+    "start": "node server/lib/server",
+    "server": "node-inspector --web-port 5858 & npm run transpile && node-dev --debug server/lib/server",
+    "test": "ava --watch",
+    "transpile": "babel server --ignore server/lib,server/static --out-dir server/lib"
   },
   "ava": {
     "failfast": true,
@@ -32,33 +34,36 @@
       ]
     }
   },
-  "author": "EDA",
+  "author": "Devon McGrath",
   "license": "ISC",
   "devDependencies": {
     "ava": "^0.15.2",
-    "babel-core": "^6.9.1",
-    "babel-loader": "^6.2.4",
-    "babel-polyfill": "^6.9.1",
-    "babel-preset-es2015": "^6.9.0",
-    "babel-preset-react": "^6.5.0",
-    "babel-preset-react-hmre": "^1.1.1",
-    "copy-webpack-plugin": "^3.0.1",
     "enzyme": "^2.3.0",
+    "json-loader": "^0.5.4",
     "node-dev": "^3.1.3",
     "node-inspector": "^0.12.8",
-    "react": "^15.1.0",
-    "react-addons-test-utils": "^15.1.0",
-    "react-dom": "^15.1.0",
-    "webpack": "^1.13.1",
+    "react-addons-test-utils": "^15.2.0",
+    "redux-thunk": "^2.1.0",
     "webpack-dev-server": "^1.14.1"
   },
   "dependencies": {
-    "babel-register": "^6.9.0",
-    "body-parser": "^1.15.1",
-    "express": "^4.13.4",
+    "babel-cli": "^6.10.1",
+    "babel-core": "^6.10.4",
+    "babel-loader": "^6.2.4",
+    "babel-polyfill": "^6.9.1",
+    "babel-preset-es2015": "^6.9.0",
+    "babel-preset-react": "^6.11.1",
+    "babel-preset-react-hmre": "^1.1.1",
+    "body-parser": "^1.15.2",
+    "copy-webpack-plugin": "^3.0.1",
+    "express": "^4.14.0",
+    "react": "^15.2.0",
+    "react-dom": "^15.2.0",
     "react-redux": "^4.4.5",
-    "react-router": "^2.5.1",
+    "react-router": "^2.5.2",
+    "react-router-redux": "^4.0.5",
     "redux": "^3.5.2",
-    "superagent": "^2.0.0"
+    "superagent": "^2.0.0",
+    "webpack": "^1.13.1"
   }
 }
