--- server/db.js
@@ -1,18 +1,0 @@
-export default {
-  getThings: getThings
-}
-
-let things = [{
-  id: 1,
-  name: 'Red thing'
-}, {
-  id: 2,
-  name: 'Blue thing'
-}, {
-  id: 3,
-  name: 'Yellow thing'
-}]
-
-function getThings () {
-  return things
-}
