--- server/index.js
@@ -1,8 +1,0 @@
-require('babel-register')({})
-
-var server = require('./server').default
-var PORT = process.env.PORT || 3000
-
-server.listen(PORT, function () {
-  console.log('Server listening on port', PORT)
-})
