--- test/helpers/setup-browser-env.js
@@ -1,3 +1,0 @@
-global.document = require('jsdom').jsdom('<body></body>')
-global.window = document.defaultView
-global.navigator = window.navigator
